Fixed all internSkills to be a hashmap of String to Double instead of String to Integer.

+++ main/Entities/InterviewIntern.java
@@ -23,7 +23,7 @@ public class InterviewIntern extends Intern{
      * @param interviewInternSkills A map of skills where the key is the name of the skill and the value is the percentage.
      */
     public InterviewIntern(String interviewInternName, int interviewInternAge,
-                           HashMap<String, Integer> interviewInternSkills){
+                           HashMap<String, Double> interviewInternSkills){
         super(interviewInternName, interviewInternAge, interviewInternSkills);
     }
